fix: 게시글 이미지 업로드 오류 및 한글 폰트 깨짐 해결
- uploadBoardImage 서버 액션을 board/_actions/로 분리
  (edit 페이지에서 new/actions.ts import 시 서버 액션 ID 불일치 오류 수정)
- 한글/특수문자 파일명 업로드 시 Content-Disposition 오류 방지 (new File로 파일명 교체)
- Server Action body 크기 제한 2mb로 상향 (next.config.ts)
- 이미지 용량 초과(1.5MB) 및 업로드 실패 시 alert 피드백 추가
- 게시글 저장/수정 실패 시 alert 피드백 추가
- S-Core Dream woff2 3종에서 blank Korean glyph 8,822개 제거
  (CSS 폰트 폴백이 동작하지 않던 근본 원인 해결)
- body font-family를 --gl-font-body로 변경하여 폴백 체인 적용
  (Apple SD Gothic Neo → Malgun Gothic → Noto Sans KR)
- AttendanceCard SSR 하이드레이션 불일치 수정 (isMounted 패턴)
- .gitignore에 docs/local/, supabase/.temp/ 추가
