+++ src/app/board/new/actions.ts
@@ -15,10 +15,6 @@ export type CreateBoardPostResult = {
   fieldErrors?: BoardPostFieldErrors;
 };
 
-export type UploadBoardImageResult =
-  | { ok: true; url: string }
-  | { ok: false; error: string };
-
 type ProfileRow = {
   last_name?: string | null;
   rank?: string | null;
@@ -46,38 +42,6 @@ function deriveAuthorName(
     lastName && rank ? `${lastName} ${rank}` : lastName || rank || null;
 
   return combined ?? '익명';
-}
-
-export async function uploadBoardImage(
-  formData: FormData
-): Promise<UploadBoardImageResult> {
-  const supabase = await createServerSupabase();
-  const {
-    data: { user },
-  } = await supabase.auth.getUser();
-
-  if (!user) return { ok: false, error: '로그인이 필요합니다.' };
-
-  const file = formData.get('file') as File | null;
-  if (!file) return { ok: false, error: '파일이 없습니다.' };
-
-  const ext = file.name.split('.').pop();
-  const fileName = `${user.id}/${Date.now()}.${ext}`;
-
-  const { error: uploadError } = await supabase.storage
-    .from('post-images')
-    .upload(fileName, file, { contentType: file.type });
-
-  if (uploadError) {
-    console.error('[uploadBoardImage] upload failed', uploadError);
-    return { ok: false, error: '이미지 업로드에 실패했습니다.' };
-  }
-
-  const {
-    data: { publicUrl },
-  } = supabase.storage.from('post-images').getPublicUrl(fileName);
-
-  return { ok: true, url: publicUrl };
 }
 
 export async function createBoardPost(
