+++ src/app/globals.css
@@ -17,7 +17,7 @@ body {
 }
 
 body {
-  font-family: var(--font-body);
+  font-family: var(--gl-font-body);
   font-size: 1rem;
   line-height: 1.625rem;
 }
@@ -73,8 +73,12 @@ body {
   color-scheme: light;
 
   /* Font variables */
-  --gl-font-brand: var(--font-brand, 'DNF Bit Bit v2'), system-ui, sans-serif;
-  --gl-font-body: var(--font-body, 'S-Core Dream'), system-ui, sans-serif;
+  --gl-font-brand:
+    var(--font-brand, 'DNF Bit Bit v2'), 'Apple SD Gothic Neo', 'Malgun Gothic',
+    'Noto Sans KR', system-ui, sans-serif;
+  --gl-font-body:
+    var(--font-body, 'S-Core Dream'), 'Apple SD Gothic Neo', 'Malgun Gothic',
+    'Noto Sans KR', system-ui, sans-serif;
 
   /* Semantic tokens - Light mode */
   --color-dark-bg: hsl(0 0% 100%);
